requirements.txt
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==6.0.0
cryptography==41.0.4
Cython==3.0.4
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.8.0
ipython==8.16.1
ipywidgets==8.1.1
Jinja2==3.1.2
jnius==1.1.0
keyring==24.2.0
langchain==0.0.321
lockfile==0.12.2
lxml==4.9.3
mock==5.1.0
numpy==1.26.1
openai==0.28.1
ordereddict==1.1
Pillow==10.1.0
protobuf==4.24.4
pyOpenSSL==23.2.0
railroad==0.5.0
simplejson==3.19.2
Sphinx==7.2.6
toml==0.10.2
tornado==6.3.3
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zipp==3.17.0
zstandard==0.21.0
grpcio==1.59.0
